Fix spelling and minor refactors
