+++ crates/emitter/src/expanded_modport.rs
